PHPLIB-81: Move CachingIterator to MongoDB\Model
